Rename blockProvidesBlockContext directive file
