+++ src/gutenberg-packages/hydration.js
@@ -2,7 +2,7 @@ import 'preact/debug';
 import { hydrate } from 'preact';
 import toVdom from './to-vdom';
 import '../directives/wp-block-type';
-import '../directives/wp-block-provides-context';
+import '../directives/wp-block-provides-block-context';
 import '../directives/wp-log';
 
 const dom = document.querySelector('.wp-site-blocks');
